Details message is truncated when pre-check fails
This commit addresses the issue of a truncated message in the details
column of the strategy-list table, which prevents the user to check
the complete error message. The fix consists in increasing the column
length.

Test Plan:

PASS: Change the strategy_step column 'details' length from 255 to
1000 and insert the values to details column with a length above 255.
Verify that the values are correctly inserted and can be verified with
'dcmanager strategy-step list'.

PASS: Deploy a system controller with 24.09.1 and a subcloud with
24.09. Cause an mgmt affecting alarm on the subcloud, in this case
250.001 (config out of date). Create a DC strategy to deploy 24.09.1
on the subcloud. Apply the strategy while subcloud is still with
250.001 alarm. Verify that the strategy deployment fails and the message
in details column is not truncated.

Closes-bug: 2080281

Change-Id: I175305dd39b560515b39e4ad31e78c7afec173f8
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>