Avoid sending first out-of-sync update during sync
During the sync process, dcorch sends two RPC notifications to
dcmanager state. One 'out-of-sync' at the beginning, and another
'in-sync' at the end of the sync process. Since these updates occur
within a short period of time, it is causing race conditions on
dcmanager state side, leading to discrepancies in the sync status for
some subcloud endpoints between dcorch and dcmanager.

This commit removes the first 'out-of-sync' update, so dcorch will
just send one RPC notification at the end of the sync process,
depending on the result of it.

Test Plan:
PASS: Manage a large number of subclouds for the first time. Verify
      dcmanager state receives only one RPC notification instead of
      two.
PASS: Verify the sync statuses are updated correctly in dcmanager
      following the result of the dcorch sync process, either
      out-of-sync or sync.
PASS: Verify there is no discrepancy between the sync statuses between
      dcorch and dcmanager.
PASS: Verify the time taken to complete the initial sync process is
      reduced to less than 3 minutes in a large lab.

Closes-bug: 2080559

Change-Id: I548069ce37d428016c6980a8c6164bd3cb5ac3a2
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>