Fix software proxy for upload-dir requests
This commit updates the USM proxy to prevent it from attempting to
upload the same files multiple times, resulting in an error. Each
release from the software API response is now correctly matched with
its corresponding file from the request, ensuring each file is
uploaded to its respective major release subdirectory only once.

Additionally, it fixes the error parser to avoid formatting software
API responses, as they do not follow the same format as cgts client
requests. This resolves the incorrect output message displayed when
attempting to delete a nonexistent release.

It also fixes the USM release URL by removing an extra '/', fixing
the issue where the proxy wouldn't work and the webob files were
not closed (flling up the /scratch storage).

Test Plan:
1. PASS - Create a directory containing multiple patches and run the
'software --os-region-name SystemController upload-dir <dir>' command.
Wait for the operation to complete successfully, then verify that the
patches were uploaded to dc-vault;
2. PASS - Upload a single patch using 'software --os-region-name
SystemController upload --local <patch>' and verify that it is
successfully uploaded to dc-vault;
3. PASS - Upload a single patch using 'software --os-region-name
SystemController upload <patch>' and verify that it is successfully
uploaded to dc-vault;
4. PASS - Attempt to delete a nonexistent release and verify that the
CLI output is: 'Error: Release <release> can not be found';
5. PASS - Upload an iso using 'software --os-region-name
SystemController upload --local <iso> <sig>' and verify that both are
successfully uploaded to dc-vault;
6. PASS - Verify that the webob file objects are properly closed
after the upload request is complete;
7. PASS - Verify that in case of an error on the software API side,
the message returned by the proxy to the CLI is in a valid format.

Closes-Bug: 2080549
Closes-Bug: 2081320
Closes-Bug: 2081321

Change-Id: I33e533854fa8f126690dda613a3b20845d490ee3
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>