Enable prestage for subcloud in apply-strategy-failed state
This commit allows subclouds in the `apply-strategy-failed` state
to initiate a prestage operation. The `apply-strategy-failed` status
can occur when a subcloud misses the prestage phase during a sw-deploy
orchestration. This update ensures that orchestration marks the
`deploy_status` of a subcloud as complete after successfully finishing
the operation and for the prestage command.

Test Plan:

PASS: Perform a sw-deploy orchestration succesfully
- deploy_status should be `complete`
PASS: Perform a prestage successfully if the subcloud deploy_status
      is in `apply-strategy-failed` state
- deploy_status should change to `complete`

Failure Path:

PASS: Perform a sw-deploy orchestration without prestage
- deploy_status should be `apply-strategy-failed`

Story: 2010676
Task: 50826

Change-Id: I73b21e4984b254bd7b9af46eedc6fd299af57f9a
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>