Enable subcloud prestage (--for-install) for 'unavailable' state
Currently to determine if the release is valid for install,
it is initially compared to the release of the system controller.
If it matches, it is assumed that the release is already deployed.
Otherwise, the query is made in the list of software for the
available status, since those releases that have not yet been
deployed must be considered.

Due to change in status in USM API for N-1 load from 'available'
to 'unavailable', this commit includes 'unavailable' status for
prestage (--for-install).

Prestage (--for-install) now allows deployed(24.09), available
(24.09.1) and unavailable(22.12).

Add release pattern validation(YY.MM) for previous release,
as software list output for previous release is YY.MM format.

Test-Plan:

Software-list
+-------------------+------+-------------+
| Release           | RR   |    State    |
+-------------------+------+-------------+
| starlingx-22.12   | True | unavailable |
| starlingx-24.09.0 | True |  deployed   |
+-------------------+------+-------------+

Setup
System Controller 24.09
Subcloud 22.12
1)Prestage is successful --for-install.

Setup
System Controller 24.09
Subcloud 24.09
1)Prestage is successful --for-install.
2)Prestage is successful --for-sw-deploy.

Closes-Bug: 2080504

Change-Id: Ib1ae79389aebe057d6d3d580516dc96a9baefa43
Signed-off-by: sbhardwa <sharang.bhardwaj@windriver.com>