Update alarm summary when subcloud becomes offline
When a subcloud is offline, the 'dcmanager alarm summary' command
shows the previous alarm information, which can be confusing. The
same issue is present in the Horizon GUI.

This commit fixes this issue by updating the subcloud_alarms table
when a subcloud becomes offline, setting all alarms to '-1' and
the status as 'disabled'. This way, the 'dcmanager alarm summary'
command will display the correct status, and a dash '-' for offline
subclouds.

Test Plan:
PASS: Run the 'dcmanager alarm summary' command with these changes.
      Verify that the output remains unchanged for online subclouds,
      and is as expected for offline ones.
PASS: Login to Horizon. Verify that the alarms are showed for online
      subclouds, and the text "N/A" is displayed for offline ones.

Closes-bug: 2080679

Change-Id: I0a6dd84f52b14027c5963f40a492ad3db31c59af
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>