Add usm to exclude list of users for identity sync
Software commands are not working on subcloud with previous rel ISO
after working for some time. System Controller and subcloud are
installed with different credentials for 'usm' user and when it gets
synced, software commands are failing in subcloud.

This review adds 'usm' user to the list of users that are excluded
for identity sync

Test Plan:
PASS: Build, install & check subcloud software commands working after
      few hours

Story: 2010676
Task: 51040

Change-Id: Idb15bac68ba61f1613704e54e0a17aff55ac4c58
Signed-off-by: Joseph Vazhappilly <joseph.vazhappillypaily@windriver.com>