Fix audit update of endpoint status to None
In dcmanager's audit, when an endpoint's audit is skipped, it returns
None as the endpoint's sync status, but this value is not filtered
before sending the update to state, which, in some cases, might lead to
an incorrect update.
This commit fixes this issue by including a new validation for the
endpoint_data dictionary before the state request.

Test plan:
1. PASS: Unamange a subcloud and verify that all endpoints are set to
   unknown.
2. PASS: Manage a subcloud and verify that all endpoints are set to
   in-sync.
3. PASS: Update a single endpoint's sync status to None and verify that
   no state requests are made.
4. PASS: Using test 4, set another endpoint to out-of-sync and verify
   that only this endpoint is sent to state for update.

Closes-Bug: 2081207

Change-Id: I692373f72d897385c6488c2e335bf63f712da33a
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>