Fix subcloud status update after enrollment playbook failure
This commit fixes a condition where the playbook fails, but subcloud
deployment continues anyway. After the exception is raised, a "False"
condition should be returned.

This commit also fixes an enroll validation failure where the validation
method should verify if the object "install-values" is received but
verifies if the string "install-values" is received instead.

Test Plan:

PASS: Deploy a system controller and add a subcloud passing --enroll
parameter and a wrong bmc-password to force a failure in the init-
enroll playbook. Verify that the subcloud state is "pre-init-enroll-
failed".

PASS: Deploy a system controller and successfully enroll a subcloud.

PASS: Enroll a subcloud using --deploy config. Verify that the
subcloud is successfully enrolled.

Closes-bug: 2079853

Change-Id: Ia14979d9d6549b28d5717a57e93ad766009b4b6a
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>