Remove needless out-of-sync notifications
The out-of-sync update notifications, one for identity and one for
platform endpoints at the beginning of a sync routine do not add value
while creating unnecessary load on the system controller. This commit
removes these unnecessary notifications.

Test Plan:
PASS: Manage a large number of subclouds for the first time. Verify
      dcmanager state receives only one RPC notification for each
      subcloud instead of two.
PASS: Verify the sync statuses are updated correctly to out-of-sync or
      in-sync in dcmanager following the result of the dcorch sync
      process.
PASS: Verify there are no discrepancies between the sync statuses of
      dcorch and dcmanager.
PASS: Verify the time taken to complete the initial sync process is
      reduced to less than 5 minutes in a large lab.
PASS: Verify that if there is an expired certificate at the time of
      managing a subcloud, its platform sync_status will be
      'out-of-sync'. The state will remain as 'out-of-sync' until the
      expired certificate is uninstalled from the SystemController,
      and the dcorch audit job creates a new sync request.
PASS: Run the manage and induce a SyncRequestFailed exception during
      the sync process. Verify that the endpoint status is set to
      'out-of-sync', and the 280.002 alarm is raised.

Closes-bug: 2080559

Change-Id: I1d1d8188a947ee8dafe2ffb060629e9e794f43fa
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>