Retry for Keystone ConnectTimeout or NotFound
A Keystone ConnectTimeout or NotFound may occur while obtaining
a token or accessing the Keystone service from a subcloud when the
network is unstable. Therefore, two retries are being added for these
circumstances, with a random delay between attempts.

Test plan:
PASS: Successfully run backup operation on batches of sublcouds
      without encountering such errors on a scale lab.

Closes-Bug: 2083075

Change-Id: Ie5d3ad8b26589000e2724df7a8c884b4ecb1be54
Signed-off-by: lzhu1 <li.zhu@windriver.com>