Add the region_name field in the API response for subcloud detail
This commit adds the region_name field to the API response when
requesting the subcloud detail, allowing backward compatibility
of the dcmanager client to display the region_name field using
the show detail subcommand.

Originally the region_name was displayed only when the subcloud's
show --detail parameter was provisioned. It then became displayed
by default as region-name instead of region_name, without the
need to provision --detail, whose change is reflected in the
following commit;

https://review.opendev.org/c/starlingx/distcloud/+/909976

This generates a compatibility issue in clients that have already
been released, where it is necessary to use the --detail parameter
to be able to display the region_name.

With this change the following is assumed:
If the API is requested to display the subcloud information, the
region name will be region-name by default.
If the API is requested to show the details of the subcloud, the
region name field will be shown with its two variants; region-name
and region_name.

Test Plan:

PASS - Send a request to the API URL to display the subcloud
       information.
     - Verify that the region name field is displayed as
       "region-name".

PASS - Send a request to the API URL to display detailed subcloud
       information.
     - Verify that the region name field is shown with its two
       variants; "region-name" and "region_name".

Closes-bug: 2083459

Change-Id: I37cde81dea3df20cbf6e5f268998bd044d5a9334
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>