Fix subcloud enroll retry failure
This commit fixes the "user is undefined" error raised
when running an enroll retry operation.

Test Plan:

PASS: Deploy a subcloud with a certificate installed
that will expire in a few minutes. Enroll the subcloud
when the certificate is expired failing the playbook.
Install a valid certificate and retry the operation.
Verify that the subcloud is enrolled successfully.

PASS: Create and bootstrap a subcloud with phased
deploy method. Verify that the subcloud is correctly
deployed.

PASS: Force a subcloud enroll failure and remove the
subcloud.yml file from dc-vault. Enroll the subcloud
again and verify that the subcloud enrolls successfully
without any errors.

Closes-bug: 2083562

Change-Id: I2e632b2706a39020f2a5d30025a4c8091d7000b8
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>