Subcloud Remote Pre-Install - Optimize Stale Mount Checks
Replaced the sh lib call with os lib to avoid spawning new
processes when checking the directory inode numbers.

These changes address a remote pre-install performance regression
observed during parallel subcloud add (100 subclouds), where
subclouds were spending excessive time (up to 17 minutes) in the
pre-install state. Overall, the average pre-install time is
reduced by 47% during parallel subcloud add.

The cause of this delay was traced to the following change:
[1] https://review.opendev.org/c/starlingx/distcloud/+/920194

In this change, the critical pre-install section was expanded to check
for stale OSTree bind mounts. Specifically, the use of the
sh module in this hot path (under a lock) was identified as a bottleneck
with batch add.

Test Plan:
1. PASS: Re-run original fix test plan [1] -
  - Unmount (but do not delete) the /var/www/pages/iso/<rel>/ostree_repo
    directory. When a subcloud add or deploy operation is done, the bind
    mount is recreated.
  - Stale mount:
      # Replace the original
      sudo cp -a /var/www/pages/feed/rel-24.09/ostree_repo \
        /var/www/pages/feed/rel-24.09/ostree_repo.orig
      sudo rm -rf /var/www/pages/feed/rel-24.09/ostree_repo
      sudo cp -a /var/www/pages/feed/rel-24.09/ostree_repo.orig \
        /var/www/pages/feed/rel-24.09/ostree_repo
    When a subcloud add or deploy operation is done, the stale bind
    mount is detected. The /var/www/pages/iso/24.09/ostree_repo is
    unmounted, and the directory is removed.
    When a subcloud add or deploy operation is done, the bind
    mount is recreated.
2. PASS: Scale remote install

Closes-Bug: 2084063

Change-Id: Ib7737d0ee03c54b47f2f4ceddf1d0088b26b410d
Signed-off-by: Salman Rana <salman.rana@windriver.com>