Avoid timeouts while getting keystone token on orchestration
During orchestration apply, some subclouds can fail because of timeout
on getting a keystone token. This happens because the timeout was
previously reduced from 15 to 5 to accommodate the audit of a large
number of offline subclouds. This commit addresses this by specifying
a new config for http_discovery_timeout with 15 seconds by default and
setting as 5 seconds only on audit workers.

Test plan:
  - Perform a kube-rootca-update orchestration in a scale environment.
    Verify no errors from keystone happen after issuing
    'dcmanager kube-rootca-update-strategy apply'
  - Verify the discovery timeout is maintained at
 5s for dcmanager-audit-worker as intended previously.
Closes-bug: 2084599

Change-Id: I79188ec83cf96319cc9ee719b5cc61637527f4e4
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>