Refresh service list at every token refresh cycle
At the first use of endpoint cache, both the master service list and
master endpoint list are cached for subsequent use. Although master
endpoint list is refreshed at every call to get the token, master
service list does not get refreshed. This will cause a problem if a
new service and endpoints are added to central cloud, as a list with
the new endpoint will be retrieved but without a match in the cached
service list. This commit fixes this issue by forcing the service list
to refresh alongside the endpoint list at every new token cycle.

Test plan:
  - PASS: Manually add a new service and endpoint with "openstack
          service create [...]" and "openstack endpoint create [...]".
          Wait an hour for the cached token to become invalid and
          verify endpoint cache was successful in building the master
          service endpoint map.

Closes-bug: 2085176

Change-Id: Id33215c2a583a55cc36afe8d5270f5c835fd6186
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>