Improve dcmanager's prestaging strategy execution
This commit improves dcmanager's prestage strategy execution by removing
unique database queries to set the subclouds' prestage_status and
prestage_versions during execution and replacing them by a bulk request
in the orchestrator's worker thread, thus reducing multiple database
queries for a single subcloud by one that updates all at once.
Additionally, it also removes the storage of the oam floating ip for the
subclouds's in the strategy's extra args in the database by an in memory
storage in each of the orchestration workers, improving the retrieval
time for the data.
Lastly, it also stops the audit execution for strategy steps that are
not in aborted, complete or failed states during an active strategy
processing.

Note: the unit tests will be completed in a subsequent review.

Test plan:
1. PASS: Create and apply a prestage strategy and verify that the steps
   are updated in bulk everytime the orchestration loop is finished and
   there is data to update.
2. PASS: In the prestage strategy, verify that the prestage_status field
   is set to 'prestaging' when the subcloud is being prestaged and that
   it is successfully updated to either complete or failed once its
   finished.
3. PASS: Create a prestage strategy, apply and abort it, verifying that
   the steps that did not start prestage do not have the prestage_status
   field filled.
4. PASS: Complete the prestage orchestration successfully for a subcloud
   and verify that the prestage_versions field is correctly set with the
   prestaged version.
5. PASS: Run a standalone prestage and verify that the prestage_status
   and prestage_versions fields are correctly set in the subcloud.
6. PASS: Try to run a standalone prestage once there is an active
   prestaging and verify that the error message correctly displays the
   prestage_status.
7. PASS: Verify that the audit execution is skipped for subclouds being
   applied or in initial state during an active strategy.
8. PASS: Delete a strategy and verify that the steps, up to max parallel
   subclouds, that are either in complete or failed state have their
   updated_at field reset before processing starts.
9. PASS: Start a prestage strategy, restart dcmanager-manager service
   and verify that the subclouds in failed state can be prestaged again
   once the strategy is recreated.

Story: 2011311
Task: 52078

Change-Id: Ib700cc662dcfc6728a8c2a4d6eb1233b1187fe44
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>