Clear dcagent cache upon failure
Currently, if a call to update dcagent cache fails, the previous
successful response will still be cached, so when dcmanager/dcorch
attempts to audit, an out-of-date response will be used. This commit
fixes this problem by clearing the cache for the function if the
call returns an error.

Test plan:
  - PASS: Manually force an error in software api. Verify the cache
          from dcagent is cleared and the next audit fails.
  - PASS: Manage a healthy subcloud and verify the audit is working
          as expected.

Closes-bug: 2085832

Change-Id: Ib97f58e8d20e6805b08e930b6e4200aae5451719
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>