Support prestage N release --for-sw-deploy on subclouds N-1 without USM
This commit removes the USM verification logic. That is, it will
not be necessary to verify if the subcloud has USM enabled,
even if the subcloud is with release N-1.

In this way, release N can be prestaged in subcloud N-1, regardless
of whether the subcloud has USM enabled or not.

Test Plan:
Setup: - Bring up a SC with 24.09.
PASS - Bring up an N-1 subcloud without USM support.
     - Prestage the subcloud for software deploy.
     - Verify that the 24.09 release is prestaged on the subcloud.

PASS - Create and apply a prestage strategy for software deploy.
     - Verify that the 24.09 release is prestaged on the subcloud.

PASS - Bring up an N-1 subcloud with USM support.
     - Prestage the subcloud for software deploy.
     - Verify that the 24.09 release is prestaged on the subcloud.

PASS - Create and apply a prestage strategy for software deploy.
     - Verify that the 24.09 release is prestaged on the subcloud.

PASS - Bring up a subcloud with 24.09.
     - Upload and deploy patch 24.09.1 on SC.
     - Prestage the subcloud for software deploy.
     - Verify that the 24.09.1 patch is prestaged on the subcloud.

PASS - Create and apply a prestage strategy for software deploy.
     - Verify that the 24.09.1 patch is prestaged on the subcloud.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/933673

Story: 2010676
Task: 51415

Change-Id: Ief595ee4193fd0a6b196d69d8c221f95d2c0dbd4
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>