Update audit new endpoint flag for software (usm)
This commit updates the audit flag to the recently added software
endpoint in the DCManager audit. This ensures the creation of the
software endpoint in the database after a system controller upgrade.

Test Plan:
PASS: Perform system controller upgrade and verify that the new
      software endpoint is added to the database successfully.
PASS: Create a sw-deploy strategy for a n-1 subcloud.

Closes-bug: 2086860

Change-Id: I8bd67fefba1f6a3d54c3b0fbd65674de1dcdbb86
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>