Improve subcloud_id retrieval in dcorch
The introduction of a subcloud_id cache in [1] caused an issue
when a subcloud is deleted from the SystemController, as the
cache was not being updated. This leads to issues when a subcloud is
rehomed back, resulting in an invalid subcloud_id during dcorch audit.

This commit removes that cache, and passes the subcloud ID during the
sync object initialization, as previosly suggested in [1]. This
approach also removes the redundant get_by_name calls.

Test Plan:
PASS: Deploy and manage a subcloud. Verify the initial sync process
      completes successfully.
PASS: Migrate it to a different SystemController, delete it, and
      rehome it back. Verify that dcorch audit and sync complete
      successfully.
PASS: Manage a batch of subclouds on a scale lab. Verify that the
      initial_sync process is completed successfully for all the
      subclouds, and that the execution time remains consistent with
      the results from [1].

Closes-bug: 2088071

[1]: https://review.opendev.org/c/starlingx/distcloud/+/927644

Change-Id: I97cb80db475adbcbab3114cdc514314f5ca3648b
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>