Synchronize subcloud software version after manual upgrade
When the subclouds are manually upgraded by directly calling VIM APIs,
the dcmanager and dcorch databases on the system controller wouldn’t
automatically detect and update the software version. This task aims to
update the software version in the dcmanager and dcorch databases
if the software audit finds a different version on the subcloud.
Additionally, the software and Kubernetes audit intervals are changed
to 40 seconds to align with the interval of periodic subcloud audit.

Test Plan:
PASS: Verify that software and kubernetes audits run every 40 seconds.
PASS: Confirm that the subcloud software version is updated in both
      the dcmanager and dcorch databases following a manual upgrade and
      that the software audit status changes from out-of-sync to
      in-sync.

Story: 2010651
Task: 51334

Change-Id: Ib4657400362c59757b9946ce79304206bc548c9b
Signed-off-by: lzhu1 <li.zhu@windriver.com>