Notify dcorch of software version change on backup restore
It's possible to restore a backup of a previous release. In the
process, dcmanager db is updated with the specified release, but not
dcorch db, which would cause a problem in audit when managing the
subcloud upon restore completion if going from previous release, as
the software version is used to determine if dcagent should be used
for platform audit and it won't be available at previous release.
This commit fixes this by updating dcorch db as well.

Test plan:
  - PASS: Perform a backup restore passing the previous release when
          the installed version is current release. Verify dcorch db
          is also updated in the process and the subcloud can be
          successfully managed at the end.

Closes-bug: 2089562

Change-Id: I2ad0a9d427ed561ececf4d0f3b195da33ad40640
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>