Fix DC geo-redundancy systemcontroller_gateway_address syncing
Previously, when the secondary site association sync status was set to
'out-of-sync', when the primary site becomes online again, it would
update the primary site rehome-data with the secondary site data. This
causes an issue because the systemcontroller_gateway_address present on
the rehome-data is exclusive to each site and can't be shared across
sites, causing an invalid route being created on the primary site.
This commit prevents that by only syncing all the other rehome-data
attributes, except the systemcontroller_gateway_address.

This commit also prevents the association sync status from being set as
'out-of-sync' if the local association is already marked as 'in-sync'
whenever the primary site detects that the secondary site is reachable.
It also causes the peer monitor thread to immediately audit the peer
site after the association is created, instead of waiting for the
heartbeat interval.

Additionally, this commit fixes an incorrect dictionary key usage from
"sync_status" to "sync-status" that caused the peer site association
sync status to be updated unnecessarily.

It also fixes an issue introduced by [1] where the subcloud update
would fail when there was an update in the
systemcontroller_gateway_address attribute because the validation
function was expecting some attributes that were not available as part
of the request payload.

This commit also improves some log messages and adds type annotations
to some geo-redundancy related methods.

Test Plan:
01. PASS - Run an end-to-end geo-redundancy test, migrating a subcloud
    from the primary site to the secondary and back.
02. PASS - Re-run the GR test, but cause a failure during the migration
    to the secondary site. Update the bootstrap values on the secondary
    site and verify that the sync-status is out-of-sync. Run the
    migration to secondary site again, then migrate back to primary
    site and verify that the rehome-data synchronization does
    not synchronize the systemcontroller_gateway_address attribute.
03. PASS: Do the following steps:
    - Create a system peer with an incorrect systemcontroller
      gateway address that's inside the management subnet, but
      outside the reserved IP range and then create an association.
      Verify that the secondary subcloud and a route was created
      using the incorrect IP.
    - Update the system peer with the correct systemcontroller
      gateway address on the primary site. Verify that the PGA
      sync status is set to 'out-of-sync' on both sites.
    - Sync the PGA and verify that the secondary subcloud
      systemcontroller gateway address was updated and that the
      old route was deleted and a new one using the new address
      was created.
    - Migrate the SPG to the non-primary site and verify that
      it completes successfully and that the subcloud becomes
      online and managed.
04. PASS - After creating a peer group and the association, verify that
    the peer monitor thread is started and that the first heartbeat
    check is executed without waiting for the heartbeat interval.

[1]: https://review.opendev.org/c/starlingx/distcloud/+/922255

Closes-Bug: 2089715

Change-Id: I857f30e2d691dfb18196f123ba5a2a52fd8ddb64
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>