Enhance logging for dcaudit and dcorch
To improve the logged information during dcmanager audit and avoid
unnecessary logging in dcorch and audit, this commit adjusts log
levels between DEBUG and INFO.
Additionally, a specific rule is added for state.log in logrotate to
retain larger files compared to the rest of the dcmanager logs.

It was also seen that at the end of the dcorch audit process, the
sync_request was being updated to 'requested' for the subcloud and
endpoint being audited. This commit enhances this logic by updating
this state just for the cases in which there are sync jobs to be run.

Test Plan:
PASS: Verify that during dcorch audit, if there are no sync jobs to be
      run, the sync_request field is not updated.
PASS: Verify that during dcorch audit, if there are sync jobs, the
      sync_request field is updated to 'requested'. Also verify that
      in the next sync round, that subcloud/enpdoint is synched.
PASS: Verify that state.log rotates when it reaches the configured
      size of 40MB.
PASS: Verif that other dcmanager logs follow the previous rotation
      policy.

Story: 2011106
Task: 51424

Change-Id: Icfa3eb9bd692e1e22f8940c904e7ef373260589d
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>