Prevent subcloud deletion during software deployment
This commit fixes the subcloud database deletion when a software
release is being applied on the controller, by preventing the
delete operation from being happening.

Test Plan:

PASS: Apply a patch on the controller and wait for its apply to
finish. Verify that the subcloud can be deleted.

PASS: Have a subcloud in an offline and unmanage state and
start deploying a patch on the system controller. During patch
application, attempt to delete subcloud and verify the expected
failure. Once the patch application completes, power on the
subcloud and attempt to manage it. Verify that the subcloud is
managed and it was not deleted from the dcorch and/or dcmanager
databases.

Closes-bug: 2089928

Change-Id: I64979210372f796d6f329c00f07726e32acb5c34
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>