Set software audit sync status to out-of-sync based on deploy_status
This commit ensures the `sync_status` for the software audit is set
to `out-of-sync` if the subcloud `deploy_status` is either
`sw-deploy-apply-strategy-failed` or `sw-deploy-in-progress`. This
prevents a scenario where the DCManager marks the status as `in-sync`
without updating the `deploy_status` if the VIM completes the release
deployment after a failed `sw-deploy` orchestration.

Additionally, this commit introduces:
- A new precheck in the `sw-deploy` orchestration to verify if the
  release is already deployed (for apply or removal).
- An increased timeout for the VIM apply strategy operation.

Test Plan:
PASS: Perform a successful `sw-deploy` orchestration.
PASS: Confirm that a failed `sw-deploy` orchestration sets the
      software to `out-of-sync` based on the `deploy_status`.
PASS: Validate that after a failed `sw-deploy` orchestration, if the
      subcloud deploys the release independently, a reapply detects
      the deployed release and updates the `deploy_status` to
      `complete`.
PASS: Verify that the if a release was removed in the subcloud
      independently/manually, an apply/reapply of the release
      removal via the orchestrator will result in the deletion
      of the release on the subcloud.
PASS: Verify that the license install step is skipped for minor
      release.
PASS: Verify that no dc audits occur during the activation phase
      of a major release. Verify that audits resume when deployment
      is complete.

Closes-bug: 2091018

Change-Id: I8775fdaeaf7b6dce81e928fddc7280e3543445f4
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>