Revert sync_request update during dcorch audit
Commit [1] introduced an error during unmanage-manage of a
subcloud, leaving the platform and identity endpoints as
'unknown'. The same issue occurs when a subcloud transitions
from offline to online.

This commit reverts the logic of updating the sync_request when
there are no audit jobs. A subsequent commit will address this
enhancement.

Test Plan:
PASS: Unmanage and manage a subcloud. Verify that the identity and
      platform endpoints are synchronized and their statuses are
      in-sync.
PASS: Verify that when a subcloud goes from offline to online, the
      identity and platform endpoints are synchronized and their
      statuses are in-sync.

Closes-bug: 2091034

[1]: https://review.opendev.org/c/starlingx/distcloud/+/936655

Change-Id: I4ff2d198f928b33d3167a0d83380afd347de95cc
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>