Prestage versions not displayed after completion
After the subcloud prestage process is complete, the prestage versions
are not displayed under the prestage_versions field. The issue comes
from an incorrect playbook task name used to search for the prestage
versions in the Ansible playbook logs.

Test Plan
PASS - Verify that the prestage_versions field displays the correct
       values after a successful prestage operation.

Closes-Bug: 2091967

Change-Id: I092aede91d315ecedd90340e84d23cbbf66eb013
Signed-off-by: lzhu1 <li.zhu@windriver.com>